Split StakePoolMetadata module out of Cardano.Api.Typed
